+++ typescript/service-worker/src/http-interface/canister_http_interface_types.d.ts
@@ -8,7 +8,7 @@ export interface HttpRequest {
   method: string;
   body: Uint8Array;
   headers: Array<HeaderField>;
-  certificate_version: [] | [bigint];
+  certificate_version: [] | [number];
 }
 export interface HttpResponse {
   body: Uint8Array;